requirements.txt
Flask==2.2.5
PyYAML==6.0.2
requests==2.31.0
SQLAlchemy==2.0.31
